+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.1.1
+- compendium with demo macros
+- you can use macros inside compendium
+- BREAKING: openCustomMacroManager use an object as arg. Read docs
+
 0.1.0
 - create custom macro, no need for settings